システム（触らないでください）/requirements.txt
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
flask>=3.0.0
anthropic>=0.40.0
playwright>=1.40.0
playwright-stealth>=1.0.6
python-dateutil>=2.8.2
requests>=2.31.0
beautifulsoup4>=4.12.0
